Added first test suite for the AppComponent
Added the complete test suite for the AppComponent, which basically
tests the router behaviour. To make all tests work, the "as:" part in
the RouterConfig in main.ts had to be changed to "name:". After this,
all specs went green.

The test suites for the other components have also been added, however,
they are still empty.

!!!Importatnt: The console output of the test.html page still shows
   errors, although the jasmine output is completely green...